Rename formatter class and move it to utils package
